backend/requirements.txt
boto3==1.4.4
Django==1.11
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
djangorestframework-queryfields==1.0.0
django-cors-headers==2.0.2
django-filter==1.0.2
django-storages==1.5.2
gunicorn==19.7.1
PyMySQL==0.7.9
PyJWT==1.4.2
pycodestyle==2.3.1
requests==2.13.0
